Kernel/aarch64: Rename AUX.{h,cpp} to AUXPeripherals.{h,cpp}
The file name "AUX" is reserved on Windows.
